migrated to multiple datafiles to separate aggregate vs nodewise data schemas
